chore: rename folder hrs to lines
