+++ frontend/src/pages/PageHome/PageHome.tsx
@@ -1,6 +1,6 @@
 import SectionAboutUs from "../../widgets/layout/sections/SectionAboutUs/SectionAboutUs";
 import SectionHero from "./ui/SectionHero/SectionHero";
-import DividerBetweenSections from "../../shared/ui/hrs/DividerBetweenSections";
+import DividerBetweenSections from "../../shared/ui/lines/DividerBetweenSections";
 import SectionProjects from "../../widgets/sections/SectionProjects/SectionProjects";
 import SliderTestimonials from "../../widgets/sliders/SliderTestimonials/SliderTestimonials";
