+++ frontend/src/shared/ui/lines/DividerBetweenSections.tsx
